Appmod/typescript upgrade 20260617223406

+++ backend/src/main/java/com/ethicalsoft/ethicalsoft_complience/application/usecase/project/ProcessExpiredProjectIsepUseCase.java
@@ -151,7 +151,7 @@ public class ProcessExpiredProjectIsepUseCase {
                             .filter(qt -> qt.getId().equals(qResult.getQuestionnaireId()))
                             .findFirst().orElse(null);
             BigDecimal weight = (q != null && q.getWeight() != null)
-                    ? BigDecimal.valueOf(q.getWeight())
+                    ? q.getWeight()
                     : BigDecimal.ONE;
             weighted.add(new IsepMath.WeightedValue(qResult.getIseq(), weight));
         }
